pyproject.toml
[tool.poetry]
name = "autorent"
version = "0.1.0"
description = ""
authors = ["aminbidad1 <amin.200.bidad@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.4"
environs = "^11.2.1"
djangorestframework = "^3.15.2"
pormalizer = "^0.1.0"
shapely = "^2.0.6"
djangorestframework-simplejwt = "^5.3.1"
pytz = "^2024.2"
requests = "^2.32.3"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
